core_tests: fix compile failure with GCC 4.8.4
Reported and tested by smooth

+++ tests/core_tests/chaingen.h
@@ -472,7 +472,7 @@ inline bool replay_events_through_core(cryptonote::core& cr, const std::vector<t
 //--------------------------------------------------------------------------
 template<class t_test_class>
 struct get_test_options {
-  const std::pair<uint8_t, uint64_t> hard_forks[1] = {std::make_pair(1, 0)};
+  const std::pair<uint8_t, uint64_t> hard_forks[1] = {std::make_pair((uint8_t)1, (uint64_t)0)};
   const cryptonote::test_options test_options = {
     hard_forks
   };
